Updated script.js and css stylesheet
Created and appended new modal and all modal elements to the index. Added for EachLoop inside of typeElement to display all pokemon types on modal.

Added 'is-visible' class to modal to allow for it to be hidden in CSS and applied window eventlisteners to close the modal (including the escape kley)